Move root entry points into `src`.
